UserService converted to interface
